requirements.txt
django
djangorestframework
djangorestframework-jwt
djangorestframework_simplejwt
psycopg2-binary
requests
drf_yasg